Upgrade vite, vitest and associated packages (#1258)
