+++ package.json
@@ -76,7 +76,7 @@
     "@types/react-simple-maps": "^3.0.4",
     "@typescript-eslint/eslint-plugin": "^8.59.0",
     "@typescript-eslint/parser": "^8.59.0",
-    "@vitejs/plugin-react": "^5.1.4",
+    "@vitejs/plugin-react": "^6.0.1",
     "cloudfront": "^0.4.1",
     "dotenv": "^17.4.2",
     "eslint": "^9.35.0",
@@ -102,9 +102,8 @@
     "ts-node": "^10.9.2",
     "typescript": "^5.9.3",
     "typescript-eslint": "^8.59.0",
-    "vite": "^7.3.2",
-    "vite-tsconfig-paths": "^6.1.1",
-    "vitest": "^4.1.4"
+    "vite": "^8.0.9",
+    "vitest": "^4.1.5"
   },
   "overrides": {
     "react-simple-maps": {
